habit_tracker&visualiser/requirements.txt
MySQL-python
pandas
matplotlib
mysql-connector-python